requirements.txt
alembic==1.4.2
astroid==2.4.2
cffi==1.14.2
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jose==1.0.0
jwt==1.0.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
